CAMEL-11884: NPE when stopping Salesforce compo...
...nent that failed to start

Adds a check to see if `config` is set in order to prevent NPE when
stopping.

+++ components/camel-salesforce/camel-salesforce-component/src/main/java/org/apache/camel/component/salesforce/SalesforceComponent.java
@@ -386,7 +386,7 @@ public class SalesforceComponent extends DefaultComponent implements VerifiableC
                 // shutdown http client connections
                 httpClient.stop();
                 // destroy http client if it was created by the component
-                if (config.getHttpClient() == null) {
+                if (config != null && config.getHttpClient() == null) {
                     httpClient.destroy();
                 }
                 httpClient = null;
